requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
exceptiongroup==1.2.2
filelock==3.15.4
identify==2.6.0
idna==3.7
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pytest==8.3.2
PyYAML==6.0.2
requests==2.32.3
tomli==2.0.1
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
virtualenv==20.26.3
